requirements.txt
anyio==3.6.1
arrow==1.2.2
beautifulsoup4==4.11.1
binaryornot==0.4.4
bs4==0.0.1
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cookiecutter==1.7.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.1
gunicorn==20.1.0
h11==0.13.0
hurry.filesize==0.9
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
manage-fastapi==1.1.1
MarkupSafe==2.1.1
poyo==0.5.0
prompt-toolkit==3.0.30
pydantic==1.9.2
python-dateutil==2.8.2
python-slugify==6.1.2
pytube==12.1.0
questionary==1.10.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
text-unidecode==1.3
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
wcwidth==0.2.5
